#425 Workaround race between nchan_auth/unsub and IP detection
The nchan_auth and nchan_unsub callbacks on the cloud server currently
update the "screen version" for online/offline status notification
purposes. This races with the IP set update code, which is started and
stopped in tandem with the websocket.

This commit changes the IP monitor loop from update-sleep to sleep-update,
which gives the client time to receive the first snapshot, and makes it
more likely that the "screen version" is up to date.